feat: allow "docs" to be served through a symlink
I am preparing a script in winglang/wing for one-click local viewing of the documentation website. As part of this, the script symlinks the development repo under `./docs`.

Out of the box, Docusaurus has some issues with symlinked docs site, so this change adds a little plugin to modify the webpack behavior (as recommended somewhere on the internets).