SEO Minor Improvements: Sitemap, canonical links (#728)
Hi, I noticed the sitemap wasn't being served. I think this should fix it by changing the location to the site directory.

I also noticed that the rel="canonical" was the old URLs. I think changing this env variable in build.yml should fix that but I'm not entirely sure.