Private api tutorial (#784)
This adds a tutorial for the private api use case. It's a basic draft at
this point.

Wasn't sure where to put it, so just added it as a blog. Perhaps there's
a better place?

notes:

- [ ] there's no `wingcli` package at the moment see
https://github.com/winglang/wing/issues/5495
- [x] no template "private-api-aws" at the moment, will create a
separate PR to add this to wing
- [x] the tutorial is not working yet, since the vpc / api / lambda
platform mapping is not in wing yet
- [x] not sure if typescript will work or not, so just removed it for
now

related to https://github.com/winglang/wing/issues/5509

---------

Co-authored-by: Hasan <45375125+hasanaburayyan@users.noreply.github.com>
Co-authored-by: Hasan <hasanaburayyan21@gmail.com>
Co-authored-by: Elad Ben-Israel <eladb@monada.co>