feat(docs): moved landing page into docusuarus with a new look and feel (#979)
Our current website is in WebFlow, I thought it might be good for consistency on the website to maybe have the website in this project? We are already using tailwind in the project for the new examples page, so making a new landing page in docusuarus should be easy enough....

Having control of the website allows to do something interesting things, like code snippets, wing console embeded etc.

This is draft at the moment whilst I work on it to get any feedback.....